Fix Snyk Vulnerabilities
Bumped gulp version
Removed some useless dependencies
Combined all gulp fcts in a single file
