--- gulpTasks/configuration.js
@@ -1,8 +1,0 @@
-module.exports = {
-    autoprefixerOptions: {
-        browsers: ['cover 90%', 'last 1 versions', 'IE 11', 'not dead'],
-    },
-    gzipOptions: {
-        append: false,
-    },
-};
