+++ package.json
@@ -18,7 +18,7 @@
     "build": "gulp docs",
     "buildpages": "bundle exec jekyll build",
     "start": "concurrently \"npm run build && bundle exec jekyll server --host=0.0.0.0\" \"gulp watch\"",
-    "lint": "csscomb -l scss/**/*.scss",
+    "lint": "csscomb scss/**/*.scss",
     "version": "gulp clean && npm run build",
     "reconstruct": "gulp clean --all && npm install && gulp",
     "lintfix": "node node_modules/tsjs --all --exclude tmp/**/*.js docs/**/*.js lib/**/*.js resources/**/*.js"
@@ -39,30 +39,23 @@
   "devDependencies": {
     "autosize": "3.0.21",
     "concurrently": "4.0.1",
-    "csscomb": "3.1.7",
+    "coveo-png-sprite": "1.0.x",
+    "csscomb": "4.2.0",
     "del": "3.0.0",
-    "gulp": "3.9.1",
+    "gulp": "4.0.0",
     "gulp-autoprefixer": "5.0.0",
     "gulp-cheerio": "0.6.3",
-    "gulp-clean-css": "3.9.3",
     "gulp-concat": "2.6.1",
-    "gulp-csscomb": "3.0.8",
-    "gulp-file-insert": "1.0.3",
     "gulp-files-to-json": "0.2.1",
     "gulp-gzip": "1.4.1",
-    "gulp-help": "1.6.1",
     "gulp-if": "2.0.2",
     "gulp-json-sort": "1.0.0",
     "gulp-rename": "1.2.2",
-    "gulp-sass": "3.1.0",
-    "gulp-sourcemaps": "2.6.2",
     "gulp-svgmin": "1.2.4",
     "gulp-uglify": "3.0.0",
     "gulp-util": "3.0.8",
     "merge-stream": "1.0.1",
-    "png-sprite": "0.1.1",
     "pre-commit": "1.2.2",
-    "require-dir": "0.3.2",
     "tsjs": "0.7.x",
     "underscore": "1.9.1",
     "underscore.string": "3.3.5"
